@@ -2,7 +2,8 @@
   <PropertyGroup>
     <RepoRootDirectory>$(MSBuildThisFileDirectory)</RepoRootDirectory>
     <BaseIntermediateOutputPath>$(RepoRootDirectory)bin\obj\$(MSBuildProjectName)\</BaseIntermediateOutputPath>
-    <LangVersion>8.0</LangVersion>
+    <LangVersion>9.0</LangVersion>
+    <Features>strict</Features>
     <Authors>SomeAnonDev</Authors>
     <Company>Committee of Zero</Company>
     <PackageVersion>0.1</PackageVersion>
