@@ -1,9 +1,8 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net5.0</TargetFramework>
     <IsPackable>false</IsPackable>
-    <LangVersion>8.0</LangVersion>
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
@@ -15,6 +14,10 @@
 
   <ItemGroup>
     <ProjectReference Include="..\NitroSharp\NitroSharp.csproj" />
+
+    <!--<ProjectReference Include="..\NitroSharp.SourceGenerators\NitroSharp.SourceGenerators.csproj"
+                    OutputItemType="Analyzer"
+                    ReferenceOutputAssembly="false" />-->
   </ItemGroup>
 
 </Project>
